feat(project): rename account to project
