--- apps/account/.env.example
@@ -1,14 +1,0 @@
-# Env Config
-NODE_ENV=production
-
-# Database Config
-MONGODB_URI=mongodb://mongo/dashboard
-
-# Transport Config
-NATS_URI=nats://nats:4222
-ACCOUNT_GRPC_URL=account:5001
-
-# Logger
-LOG_USE_FILE=false
-LOG_FILE=logs/account.log
-LOG_LEVEL=verbose
