--- apps/account/package.json
@@ -1,47 +1,0 @@
-{
-  "name": "account-service",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "directories": {
-    "test": "test"
-  },
-  "dependencies": {
-    "@grpc/grpc-js": "^1.13.0",
-    "@nestjs-plugins/nestjs-nats-jetstream-transport": "^2.2.6",
-    "@nestjs/common": "^11.0.11",
-    "@nestjs/config": "^4.0.1",
-    "@nestjs/core": "^11.0.11",
-    "@nestjs/microservices": "^11.0.11",
-    "@nestjs/mongoose": "^11.0.1",
-    "joi": "^17.13.3"
-  },
-  "devDependencies": {
-    "@nestjs/cli": "^11.0.5",
-    "@nestjs/schematics": "^11.0.2",
-    "@nestjs/testing": "^11.0.11",
-    "@types/cookie-parser": "^1.4.8",
-    "@types/express": "^5.0.0",
-    "@types/jest": "^29.5.14",
-    "@types/node": "^22.13.10",
-    "@types/supertest": "^6.0.2",
-    "@typescript-eslint/eslint-plugin": "^8.26.1",
-    "@typescript-eslint/parser": "^8.26.1",
-    "eslint": "^9.22.0",
-    "eslint-config-prettier": "^10.1.1",
-    "eslint-plugin-prettier": "^5.2.3",
-    "jest": "^29.7.0",
-    "pino-pretty": "^13.0.0",
-    "prettier": "^3.5.3",
-    "source-map-support": "^0.5.21",
-    "supertest": "^7.0.0",
-    "ts-jest": "^29.2.6",
-    "ts-loader": "^9.5.2",
-    "ts-node": "^10.9.2",
-    "tsconfig-paths": "^4.2.0",
-    "typescript": "^5.8.2"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC"
-}
