+++ apps/auth/.env.example
@@ -11,5 +11,5 @@ AUTH_GRPC_URL=auth:5001
 
 # Logger
 LOG_USE_FILE=false
-LOG_FILE=logs/account.log
+LOG_FILE=logs/project.log
 LOG_LEVEL=verbose
