+++ apps/channel/.env.example
@@ -8,5 +8,5 @@ MONGODB_URI=mongodb://mongo/dashboard
 NATS_URI=nats://nats:4222
 AUTH_GRPC_URL=auth:5001
 CHANNEL_GRPC_URL=channel:5001
-ACCOUNT_GRPC_URL=account:5001
+PROJECT_GRPC_URL=project:5001
 USER_GRPC_URL=user:5001
