+++ apps/project/.env.example
@@ -1,0 +1,14 @@
+# Env Config
+NODE_ENV=production
+
+# Database Config
+MONGODB_URI=mongodb://mongo/dashboard
+
+# Transport Config
+NATS_URI=nats://nats:4222
+PROJECT_GRPC_URL=project-service:5001
+
+# Logger
+LOG_USE_FILE=false
+LOG_FILE=logs/project.log
+LOG_LEVEL=verbose
