+++ apps/project/src/infrastructure/command-client/nats/user-nats.service.ts
