+++ apps/user/.env.example
@@ -11,5 +11,5 @@ USER_GRPC_URL=user:5001
 
 # Logger
 LOG_USE_FILE=false
-LOG_FILE=logs/account.log
+LOG_FILE=logs/user.log
 LOG_LEVEL=verbose
