--- libs/common/src/grpc/proto/account.proto
@@ -1,37 +1,0 @@
-syntax = "proto3";
-
-package grpc_account;
-
-service AccountService {
-  rpc GetAccountById(GetAccountByIdRequest) returns (AccountMessageResponse) {}
-  rpc GetAccountByEmail(GetAccountByEmailRequest)
-      returns (AccountMessageResponse) {}
-  rpc AccountExists(AccountExistsRequest) returns (AccountExistsResponse) {}
-}
-
-message GetAccountByIdRequest {
-  string id = 1;
-}
-
-message GetAccountByEmailRequest {
-  string email = 1;
-}
-
-message AccountMessageResponse {
-  AccountMessage account = 1;
-}
-
-message AccountMessage {
-  string id = 2;
-  string createdAt = 3;
-  string updatedAt = 4;
-  string email = 5;
-}
-
-message AccountExistsRequest {
-  string email = 1;
-}
-
-message AccountExistsResponse {
-  bool exists = 1;
-}
