+++ nest-cli.json
@@ -1,20 +1,20 @@
 {
   "$schema": "https://json.schemastore.org/nest-cli",
   "collection": "@nestjs/schematics",
-  "sourceRoot": "apps/account/src",
+  "sourceRoot": "apps/gateway/src",
   "compilerOptions": {
     "deleteOutDir": true,
     "webpack": true,
-    "tsConfigPath": "apps/account/tsconfig.app.json"
+    "tsConfigPath": "apps/gateway/tsconfig.app.json"
   },
   "projects": {
-    "account-service": {
+    "project-service": {
       "type": "application",
-      "root": "apps/account",
+      "root": "apps/project",
       "entryFile": "main",
-      "sourceRoot": "apps/account/src",
+      "sourceRoot": "apps/project/src",
       "compilerOptions": {
-        "tsConfigPath": "apps/account/tsconfig.app.json"
+        "tsConfigPath": "apps/project/tsconfig.app.json"
       }
     },
     "admin-common": {
